Create Examples.kt
Code examples in Kotlin
This file contains some very basic Kotlin code constructs.
It is part of the JetBrains Academy Kotlin Course.
The code below are some examples of how basic Kotlin code can be written.
It is a playground with some funny implementations and some experiments on what works and what doesn't.
It is also a playground on how to write short code, how to write readable code, and how to write understandable code.
Short code is usually good. But sometimes the shortest code is not the best (and sometimes it is).